pyproject.toml
[project]
name = "ledger-event-sourcing"
version = "0.1.0"
description = "The Ledger — Weeks 9-10 event-sourcing starter"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0,<0.116",
    "uvicorn[standard]>=0.32.0,<0.35",
    "httpx>=0.27.0",
    "asyncpg>=0.29.0,<0.30",
    "anthropic>=0.40.0",
    "pydantic>=2.6.0,<3.0",
    "langgraph>=0.1.0,<0.3",
    "faker>=24.0.0,<26.0",
    "reportlab>=4.2.0,<5.0",
    "openpyxl>=3.1.0,<4.0",
    "fastmcp>=2.13.0",
    "python-dotenv>=1.0.0,<2.0",
    "pytest>=8.0.0,<9.0",
    "pytest-asyncio>=0.23.0,<0.25",
]

[project.optional-dependencies]
docling = ["docling>=2.0"]
mineru = ["mineru>=1.0"]

[tool.uv]
dev-dependencies = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
